wdqs - use the same deployment strategy on both eqiad and codfw
We tried to introduce management of config file only on cdfw to ease the
transition to data partition moved on /srv. This seem to be more trouble
than is worth it, especially since this change is only for a transition
period.

This change will go back to having a single deployment strategy and having
the location of the data partition configured via a template.

Bug: T144380
Change-Id: I5e68dd7289c50fc9af98ab3a373f93aab103f571