wdqs - generate RWStore cofig file in /etc/wdqs
Scap cannot generate configuration files in the application directory
since this directory is symlinked during the deployment. Let's move the
config to /etc/wdqs where it should be possible to generate it.

This change needs to be deployed at the same time as the related puppet
change: https://gerrit.wikimedia.org/r/#/c/310308/

Bug: T144380
Change-Id: I888ce9f5603c44900a6ead444d939a2ae2124f0b