Handle all possibly services relying on these artifacts
Restarting only wdqs-blazegraph is unfortunately rarely enough and
require further operations to restart other services:
- wdqs-updater
- wdqs-categories

And for wcqs:
- wcqs-blazegraph
- wcqs-updater

Scap can use a comma separated list of services to restart.
We can populate this list with all the services we might want to restart
and use require_valid_service: True to ignore missing services.

Bug: T221709
Change-Id: I8da162077fa7c2c388c859818e37c89f719dc3f3